pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Gendiff script to find difference between two files"
authors = ["artem <artem_smirnov88@mail.ru>"]
packages = [
    { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.11"
PyYAML = "6.0.2"

[tool.poetry.dev-dependencies]
coverage = "7.6.7"
flake8 = "7.1.1"
iniconfig = "2.0.0"
mccabe = "0.7.0"
packaging = "24.2"
pluggy = "1.5.0"
pycodestyle = "2.12.1"
pyflakes = "3.2.0"
pytest = "8.3.3"
pytest-cov = "6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
